Modify various files and add tabs
Modified home and menu files so that they both append a page div
that can be cleared. Added a tabs file that contains the home,
menu and contact tabs which can be clicked to navigate website.
Added functionality which clears and renders the appropiate page
when clicking on the respective tab. Changed mode to development
and added a devtool in webpack configuration.